pyproject.toml
[tool.poetry]
name = "repl_python3_LTTwitch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.1"
twitchio = "^1.2.3"
flask = "^2.0.1"
discord = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
